// src/interfaces/stripe.ts
export interface ISaleInfo {
  email: string;
  firstNameOnCard: string;
  lastNameOnCard: string;
  card: string; // card number
  ccExpMonth: string;
  ccExpYear: string;
  chargeAmount: number;
  cvc: string;
  uuid: string;
}

export class SaleInfo implements ISaleInfo {
  constructor(
    public email: string,
    public firstNameOnCard: string,
    public lastNameOnCard: string,
    public card: string,
    public ccExpMonth: string,
    public ccExpYear: string,
    public cvc: string,
    public chargeAmount: number,
    public uuid: string
  ) {}
}
// ...
export interface IStripeSourceReference {
  id: string;
  brand: string;
  country: string;
  last4: string;
  expMonth: number;
  expYear: number;
}
export class StripeSourceReference implements IStripeSourceReference {
  constructor(
    public id: string,
    public brand: string,
    public country: string,
    public last4: string,
    public expMonth: number,
    public expYear: number
  ) {}
}

export interface IStripeChargeReference {
  id: string;
  amount: number;
  created: Date;
  customer: IStripeCustomerReference;
  description: string;
  paid: boolean;
  source: IStripeSourceReference;
  status: string;
}
export class StripeChargeReference implements IStripeChargeReference {
  constructor(
    public id: string,
    public amount: number,
    public created: Date,
    public customer: IStripeCustomerReference,
    public description: string,
    public paid: boolean,
    public source: IStripeSourceReference,
    public status: string
  ) {}
}

export interface IStripeSubscriptionReference {
  id: string;
  userSubscriptionId: string;
  start: Date;
  end: Date;
  plan: IStripePlanSummary;
}
export class StripeSubscriptionReference implements IStripeSubscriptionReference {
  constructor(
    public id: string,
    public start: Date,
    public end: Date,
    public plan: IStripePlanSummary,
    public userSubscriptionId: string
  ) {}
}

export interface IStripePlanSummary {
  amount: number;
  id: string;
  product: string;
  interval: string;
  intervalCount: number;
  nickname: string;
}
// ...
export interface IStripeCustomerReference {
  id: string;
  email: string;
}
export class StripeCustomerReference implements IStripeCustomerReference {
  constructor(public id: string, public email: string) {}
}

export interface IStripeCustomerInvoiceReference {
  eventId: string;
  customerId: string;
  chargeId: string;
  invoiceId: string;
  subscription: IStripeSubscriptionReference;
}
export class StripeCustomerInvoiceReference implements IStripeCustomerInvoiceReference {
  constructor(
    public eventId: string,
    public customerId: string,
    public chargeId: string,
    public invoiceId: string,
    public subscription: IStripeSubscriptionReference
  ) {}
}